[JSON] use absl::variant for JSON values (#32808)
This changes the `JSON` class to use `absl::variant<>` internally for
storing the type and value.

The only API-visible change here is that now the accessor methods will
crash if you use them on the wrong type, courtesy of `absl::get<>`. This
does not appear to break any tests, so it appears that we have already
cleaned up all of the cases we had in the past where we might have done
something like that.